Extract structs/objects to separate files
